Refactoring logic into dialog, no dialog and common files
No functionality changes just refactor for ease of diff viewing
